testutils: move build.go into a subpackage
This removes a dependency on go/build from the main binary.
